requirements.txt
-e git+https://github.com/django/django.git@1.6b4#egg=django
djangorestframework==2.3.8
-e git+https://github.com/jezdez/django_compressor.git@9cf68a04e473b0bccf7e#egg=django_compressor
